+++ src/components/auth/RegisterForm/index.scss
@@ -5,7 +5,7 @@
     width: 60px;
     position: absolute;
     right: 0;
-    top: 40px;
+    top: 35px;
     margin-right: 5px;
     z-index: 999;
   }
