+++ src/components/modal/PasswordFindModal.tsx
@@ -1,6 +1,5 @@
 import type { FC } from 'react';
 import { useForm } from 'react-hook-form';
-import Input from '@/components/auth/Input/Input';
 import { useMutation } from '@tanstack/react-query';
 import { postPasswordFind } from '@/api/auth/authAPI';
 import axios from 'axios';
@@ -8,6 +7,7 @@ import './modal.scss';
 import { type EmailSchema, emailSchema } from '@/constants/schema/emailSchema';
 import { yupResolver } from '@hookform/resolvers/yup';
 import Button from '@/components/common/Button';
+import CommonInput from '@/components/common/CommonInput';
 
 interface Props {
   onClose: () => void;
@@ -42,7 +42,7 @@ const PasswordFindModal: FC<Props> = ({ onClose }: Props) => {
     <div className="modal-container">
       <form onSubmit={handleSubmit(onSubmit)} className="modal-content">
         <p>비밀번호 변경 링크를 받을 이메일을 입력해주세요.</p>
-        <Input
+        <CommonInput
           {...register('email')}
           type="emil"
           label=""
