+++ src/components/mypage/UserInfo/PasswordInfo.tsx
@@ -7,7 +7,7 @@ import { passwordChangeSchema, PasswordChangeSchema } from '@/constants/schema/p
 import { useMutation } from '@tanstack/react-query';
 import { patchUserInfo } from '@/api/auth/authAPI';
 import axios from 'axios';
-import PasswordInput from '@/components/auth/Input/PasswordInput';
+import CommonInput from '@/components/common/CommonInput';
 
 interface Props {
   userId: number;
@@ -59,17 +59,19 @@ const PasswordInfo: FC<Props> = ({ userId }) => {
       </div>
       {isEdit ? (
         <div className="userInfo-container-content">
-          <PasswordInput
+          <CommonInput
             {...register('newPassword')}
             label="비밀번호"
             placeholder="새로운 비밀번호를 입력하세요."
             error={errors.newPassword?.message}
+            isVisible
           />
-          <PasswordInput
+          <CommonInput
             {...register('newPasswordCheck')}
             label="비밀번호 확인"
             placeholder="비밀번호를 확인해주세요."
             error={errors.newPasswordCheck?.message}
+            isVisible
           />
         </div>
       ) : null}
