+++ src/styles/main.scss
@@ -33,6 +33,7 @@
 @import 'components/commentContainer';
 @import 'components/calendar';
 @import 'components/todoList';
+@import 'components/commonInput';
 
 /* 레이아웃 */
 @import 'layouts/grids';
